modified:   app.js modified:   public/logic.js new file:   public/modules/login.js new file:   public/modules/user.js modified:   public/stylesheets/style.css
